globalstyling-adding dark mode, fixed folder structure, separeting container/prezentationa component
