+++ src/assets/images/index.js
@@ -1,0 +1,1 @@
+export const pokeBall= require("./pokeball.png")
