+++ src/components/common/BaseBtn.js
@@ -1,17 +1,11 @@
 import {Button} from '@mui/material'
 
-
 export default function BaseBtn({ text, onClick, href, ...restProps }) {
 
     return(
         <Button 
             variant="outlined" 
-            sx={{
-                borderColor:"#003049", 
-                color:"#003049", 
-                ":hover":{borderColor:"#003049"}, 
-                ...restProps
-             }}
+            sx={{...restProps}}
             onClick={onClick} 
         >
             {text}
